Move tests into separate files
Move std and nostd tests into individual files to make lib.rs smaller.
