+++ src/tests/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(not(feature = "nostd"))]
+mod test_std;
+
+#[cfg(feature = "nostd")]
+mod test_nostd;
